requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.0
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
eventlet==0.33.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
platformdirs==4.3.8
PyMySQL==1.1.1
python-engineio==4.7.0
python-socketio==5.9.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
virtualenv==20.31.2
Werkzeug==3.1.3
wsproto==1.2.0